python_app/requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
openai==2.14.0
pydantic==2.12.5
pywebview==6.1
pyinstaller==6.17.0
